Reorganize file structure and rename files for consistency
