+++ src/components/Admin/Forms/InputFieldHeader.tsx
@@ -1,5 +1,7 @@
-import { AdminHelperText } from '../../HelperText';
+import { useLocation } from 'react-router-dom';
+import { AdminHelperText } from './HelperText';
 import { FaAsterisk } from 'react-icons/fa';
+import clsx from 'clsx';
 
 interface AdminInputFieldHeaderProps {
 	title: string;
@@ -12,13 +14,21 @@ export const AdminInputFieldHeader: React.FC<AdminInputFieldHeaderProps> = ({
 	required,
 	helper
 }) => {
+	const location = useLocation();
+
+	const isPartnerAddPage = location.pathname.includes('/partners/add');
+
 	return (
 		<div className='flex flex-col gap-2 font-roboto text-white'>
 			<div className='flex gap-1'>
 				<p className='text-base font-bold md:text-xl'>{title}</p>
 
 				{required && (
-					<div className='text-primary size-2'>
+					<div
+						className={clsx(
+							'size-2',
+							isPartnerAddPage ? 'text-secondary' : 'text-primary'
+						)}>
 						<FaAsterisk className='size-full' />
 					</div>
 				)}
